chore: move device-scans operation to user
