+++ ui/user/src/routes/device-scans/+page.svelte
@@ -1,0 +1,2 @@
+<script lang="ts">
+</script>
